requirements-fixed.txt
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.8
aiosignal==1.3.2
annotated-types==0.7.0
ansicon==1.89.0
anthropic==0.37.1
anyio==4.9.0
astor==0.8.1
asttokens==3.0.0
attrs==25.3.0
blessed==1.21.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
editor==1.6.6
executing==2.2.0
filelock==3.18.0
fonttools==4.58.1
frozenlist==1.6.2
fsspec==2025.5.1
geojson==3.2.0
geopandas==1.1.0
git-python==1.0.3
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.6
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.62.3
h11==0.16.0
html2image==2.0.7
html2text==2024.2.26
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.32.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
inquirer==3.4.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jinxed==1.3.0
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
lazy_loader==0.4
litellm==1.72.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.4.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.2.6
open-interpreter==0.4.3
openai==1.84.0
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.8
psutil==5.9.8
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyMuPDF==1.26.0
pyogrio==0.11.0
pyparsing==3.2.3
PyPDF2==3.0.1
pyperclip==1.9.0
pyproj==3.7.1
pyreadline3==3.5.4
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
rasterio==1.4.3
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
runs==1.2.2
scikit-image==0.25.2
scipy==1.15.3
selenium==4.33.0
Send2Trash==1.8.3
shapely==2.1.1
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
starlette==0.37.2
termcolor==2.3.0
tifffile==2025.6.1
tiktoken==0.7.0
tokenizers==0.21.1
tokentrim==0.1.13
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typer==0.12.5
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wget==3.2
wsproto==1.2.0
xmod==1.8.1
yarl==1.20.0
yaspin==3.1.0
zipp==3.22.0
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.8
aiosignal==1.3.2
annotated-types==0.7.0
ansicon==1.89.0
anthropic==0.37.1
anyio==4.9.0
astor==0.8.1
asttokens==3.0.0
attrs==25.3.0
blessed==1.21.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
editor==1.6.6
executing==2.2.0
filelock==3.18.0
fonttools==4.58.1
frozenlist==1.6.2
fsspec==2025.5.1
geojson==3.2.0
geopandas==1.1.0
git-python==1.0.3
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.6
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.62.3
h11==0.16.0
html2image==2.0.7
html2text==2024.2.26
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.32.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
inquirer==3.4.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jinxed==1.3.0
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
lazy_loader==0.4
litellm==1.72.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.4.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.2.6
open-interpreter==0.4.3
openai==1.84.0
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.8
psutil==5.9.8
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyMuPDF==1.26.0
pyogrio==0.11.0
pyparsing==3.2.3
PyPDF2==3.0.1
pyperclip==1.9.0
pyproj==3.7.1
pyreadline3==3.5.4
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
rasterio==1.4.3
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
runs==1.2.2
scikit-image==0.25.2
scipy==1.15.3
selenium==4.33.0
Send2Trash==1.8.3
shapely==2.1.1
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
starlette==0.37.2
termcolor==2.3.0
tifffile==2025.6.1
tiktoken==0.7.0
tokenizers==0.21.1
tokentrim==0.1.13
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typer==0.12.5
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wget==3.2
wsproto==1.2.0
xmod==1.8.1
yarl==1.20.0
yaspin==3.1.0
zipp==3.22.0
overpy==0.6
geopy==2.4.0
